+++ example/i2c_mpu6050.dart
@@ -6,7 +6,9 @@ import 'package:dart_periphery/dart_periphery.dart';
 import 'dart:io';
 
 // MPU-6050 Six-Axis (Gyro + Accelerometer)
+//
 // Datasheet: https://invensense.tdk.com/products/motion-tracking/6-axis/mpu-6050/
+//
 void main() {
   // Select the right I2C bus number /dev/i2c-?
   // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)