+++ lib/dart_periphery.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/cpu_architecture.dart'
-    show Uname, CpuArchitecture, CpuArch, nativeUname;
 export 'src/dummy.dart' show DummyDev;
 export 'src/errno.dart' show ERRNO, Errno, ErrnoNotFound;
 export 'src/gpio.dart'
@@ -69,9 +67,13 @@ export 'src/hardware/sht31.dart'
     show
         SHT31,
         SHT31result,
-        SHT31excpetion,
+        SHT31exception,
         sht31AlternativeI2Caddress,
         sht31DefaultI2Caddress;
+export 'src/hardware/mlx90615.dart'
+    show MLX90615, MLX90615result, MLX90615exception, mlx90615DefaultI2Caddress;
+export 'src/hardware/mcp9808.dart'
+    show MCP9808, MCP9808result, MCP9808exception, mcp9808DefaultI2Caddress;
 export 'src/hardware/utils/byte_buffer.dart' show BitOrder;
 export 'src/hardware/utils/uint.dart';
 export 'src/i2c.dart'
@@ -86,8 +88,6 @@ export 'src/i2c.dart'
 export 'src/led.dart' show Led, LedException, LedErrorCode;
 export 'src/library.dart'
     show
-        setCPUarchitecture,
-        useSharedLibray,
         useSharedLibrary,
         setCustomLibrary,
         useLocalLibrary,
@@ -97,6 +97,7 @@ export 'src/library.dart'
         setTempDirectory,
         reuseTmpFileLibrary,
         loadLibFromFlutterAssetDir,
+        loadPeripheryLib,
         getPeripheryLibPath;
 export 'src/mmio.dart' show MMIO, MMIOexception, MMIOerrorCode;
 export 'src/pwm.dart' show PWM, PWMexception, Polarity, PWMerrorCode;