+++ lib/src/library.dart
@@ -7,15 +7,23 @@ import 'dart:ffi';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:dart_periphery/src/cpu_architecture.dart';
-
 import 'native/lib_base64.dart';
 
 const pkgName = 'dart_periphery';
 
-const String version = '1.0.0';
-
 final String sharedLib = 'libperiphery.so';
+
+extension ABIParsing on Abi {
+  String getCPUarch() {
+    var arch = Abi.current().toString();
+    if (!arch.startsWith('linux')) {
+      throw LibraryException(
+          LibraryErrorCode.invalidParameter, "Not supported CPU architecture");
+    }
+    // e.g. linux_arm
+    return arch.substring(6).toLowerCase();
+  }
+}
 
 late DynamicLibrary _peripheryLib;
 bool _isPeripheryLibLoaded = false;
@@ -34,10 +42,10 @@ class PlatformException implements Exception {
 
 // Fix typo
 // https://github.com/pezi/dart_periphery/pull/20
-@Deprecated("Fix typo in method name")
-void useSharedLibray() {
-  _peripheryLibPath = sharedLib;
-}
+//@Deprecated("Fix typo in method name")
+//void useSharedLibray() {
+//  _peripheryLibPath = sharedLib;
+//}
 
 /// dart_periphery loads the shared library.
 /// See [native-libraries](https://pub.dev/packages/dart_periphery#native-libraries) for details.
@@ -49,18 +57,6 @@ void useSharedLibrary() {
 /// See [native-libraries](https://pub.dev/packages/dart_periphery#native-libraries) for details.
 void setCustomLibrary(String absolutePath) {
   _peripheryLibPath = absolutePath;
-}
-
-/// Bypasses the autodetection of the CPU architecture.
-void setCPUarchitecture(CpuArchitecture arch) {
-  if (arch == CpuArchitecture.notSupported ||
-      arch == CpuArchitecture.undefined) {
-    throw LibraryException(
-        LibraryErrorCode.invalidParameter, "Invalid parameter");
-  }
-  var cpu = arch.toString();
-  cpu = cpu.substring(cpu.indexOf(".") + 1).toLowerCase();
-  _libraryFileName = 'libperiphery_$cpu.so';
 }
 
 /// Sets the tmp directory for the extraction of the libperiphery.so file.
@@ -78,38 +74,18 @@ void loadLibFromFlutterAssetDir(bool use) {
   _useFlutterPiAssetDir = use;
 }
 
-String _autoDetectCPUarch() {
-  CpuArch arch = CpuArch();
-  if (arch.cpuArch == CpuArchitecture.notSupported) {
-    throw LibraryException(LibraryErrorCode.cpuArchDetectionFailed,
-        "Unable to detect CPU architecture, found '${arch.machine}' . Use 'setCustomLibrary(String absolutePath)' - see documentation https://github.com/pezi/dart_periphery, or create an issue https://github.com/pezi/dart_periphery/issues");
-  }
-  var cpu = arch.cpuArch.toString();
-  cpu = cpu.substring(cpu.indexOf(".") + 1).toLowerCase();
-  return 'libperiphery_$cpu.so';
-}
-
 /// dart_periphery loads the library from the actual directory.
 /// See [native-libraries](https://pub.dev/packages/dart_periphery#native-libraries) for details.
-void useLocalLibrary([CpuArchitecture arch = CpuArchitecture.undefined]) {
-  if (arch == CpuArchitecture.undefined) {
-    _peripheryLibPath = './${_autoDetectCPUarch()}';
-  } else {
-    if (arch == CpuArchitecture.notSupported) {
-      throw LibraryException(
-          LibraryErrorCode.invalidParameter, "Invalid parameter");
-    }
-    var cpu = arch.toString();
-    cpu = cpu.substring(cpu.indexOf(".") + 1).toLowerCase();
-    _peripheryLibPath = './libperiphery_$cpu.so';
-  }
+void useLocalLibrary() {
+  _peripheryLibPath = './libperiphery_${Abi.current().getCPUarch()}.so';
 }
 
 enum LibraryErrorCode {
   libraryNotFound,
   cpuArchDetectionFailed,
   invalidParameter,
-  invalidTmpDirectory
+  invalidTmpDirectory,
+  errorWritingLib
 }
 
 /// Library exception
@@ -134,9 +110,9 @@ bool isFlutterPiEnv() {
 
 var _flutterPiArgs = <String>[];
 
-/// Returns the PID of the running flutter-pi program, -1 for all other platforms.
+/// Returns the PID of the running program for linux, -1 for all other platforms.
 int getPID() {
-  if (!isFlutterPiEnv()) {
+  if (!Platform.isLinux) {
     return -1;
   }
   final dylib = DynamicLibrary.open('libc.so.6');
@@ -166,9 +142,21 @@ List<String> getFlutterPiArgs() {
   return List.unmodifiable(_flutterPiArgs);
 }
 
+void saveLibrary(File file, String base64EncodedLib) {
+  try {
+    file.createSync(recursive: false);
+    final decodedBytes = base64Decode(base64EncodedLib);
+    // hint: a crash occurs, if an used lib is written again
+    file.writeAsBytesSync(decodedBytes);
+  } on Error catch (e) {
+    throw LibraryException(LibraryErrorCode.errorWritingLib, e.toString());
+  } on Exception catch (e) {
+    throw LibraryException(LibraryErrorCode.errorWritingLib, e.toString());
+  }
+}
+
 /// Loads the Linux/CPU specific periphery library as a DynamicLibrary.
 DynamicLibrary loadPeripheryLib() {
-
   if (_isPeripheryLibLoaded) {
     return _peripheryLib;
   }
@@ -176,6 +164,8 @@ DynamicLibrary loadPeripheryLib() {
   if (!Platform.isLinux) {
     throw PlatformException('dart_periphery is only supported for Linux!');
   }
+
+  var abi = Abi.current();
 
   if (_peripheryLibPath.isEmpty) {
     if (isFlutterPiEnv() && _useFlutterPiAssetDir) {
@@ -202,35 +192,35 @@ DynamicLibrary loadPeripheryLib() {
       if (!dir.endsWith('/')) {
         dir += '/';
       }
-      if (_libraryFileName.isEmpty) {
-        _libraryFileName = _autoDetectCPUarch();
-      }
-      _peripheryLibPath = dir + _libraryFileName;
+
+      _peripheryLibPath = "${dir}libperiphery_${abi.getCPUarch()}.so";
     } else {
       // store the appropriate in the system temp directory
 
       String base64EncodedLib = '';
-      CpuArch arch = CpuArch();
-      switch (arch.cpuArch) {
-        case CpuArchitecture.arm:
-          base64EncodedLib = arm;
-          break;
-        case CpuArchitecture.arm64:
-          base64EncodedLib = arm64;
-          break;
-        case CpuArchitecture.x86:
-          base64EncodedLib = x86;
-          break;
-        case CpuArchitecture.x86_64:
-          base64EncodedLib = x86_64;
-          break;
-        default:
-          throw LibraryException(LibraryErrorCode.invalidParameter,
-              "Not supported CPU architecture");
-      }
 
       if (_libraryFileName.isEmpty) {
-        _libraryFileName = _autoDetectCPUarch();
+        switch (abi) {
+          case Abi.linuxArm:
+            base64EncodedLib = arm;
+            break;
+          case Abi.linuxArm64:
+            base64EncodedLib = arm64;
+            break;
+          case Abi.linuxIA32:
+            base64EncodedLib = ia32;
+            break;
+          case Abi.linuxX64:
+            base64EncodedLib = x64;
+            break;
+          case Abi.linuxRiscv64:
+            base64EncodedLib = riscv64;
+            break;
+          default:
+            throw LibraryException(LibraryErrorCode.invalidParameter,
+                "Not supported CPU architecture");
+        }
+        _libraryFileName = "libperiphery_${abi.getCPUarch()}.so";
       }
 
       if (_tmpDirectory.isEmpty) {
@@ -242,15 +232,24 @@ DynamicLibrary loadPeripheryLib() {
             "Temp directory does not exist");
       }
 
-      String path = _tmpDirectory + Platform.pathSeparator + _libraryFileName;
-      final file = File(path);
-      if (!file.existsSync() || !_reuseTmpFileLibrary) {
-        file.createSync(recursive: false);
-        final decodedBytes = base64Decode(base64EncodedLib);
-        file.writeAsBytesSync(decodedBytes);
+      if (_reuseTmpFileLibrary) {
+        String path = _tmpDirectory + Platform.pathSeparator + _libraryFileName;
+        final file = File(path);
+        if (!file.existsSync()) {
+          saveLibrary(file, base64EncodedLib);
+        }
+        _peripheryLibPath = path;
+      } else {
+        // fix https://github.com/pezi/flutter-pi-sensor-tester/issues/1
+        String path =
+            '$_tmpDirectory${Platform.pathSeparator}pid_${getPID()}_$_libraryFileName';
+        final file = File(path);
+        // avoid crash writing this file again from an other isolate
+        if (!file.existsSync()) {
+          saveLibrary(file, base64EncodedLib);
+        }
+        _peripheryLibPath = path;
       }
-
-      _peripheryLibPath = path;
     }
   }
 
@@ -260,7 +259,8 @@ DynamicLibrary loadPeripheryLib() {
   return _peripheryLib;
 }
 
-/// Returns the path of the periphery library. Empty string if the library is not loaded.
+/// Returns the path of the periphery library. Empty string. if the
+/// library is not loaded.
 String getPeripheryLibPath() {
   return _peripheryLibPath;
 }